requirements.txt
fastapi[standard]
SQLAlchemy
psycopg2
python-dotenv
alembic